theme switcher dropdown done
