+++ components/NavBar.tsx
@@ -1,8 +1,7 @@
 import React from "react";
-import ColorSwitcher from "./ColorSwitcher";
 import { useTheme } from "next-themes";
 import Logo from "./Logo";
-import ColorSwitcher2 from "./ColorSwitcher2";
+import ThemeSwitcher from "./ThemeSwitcher";
 
 const NavBar = () => {
   const { resolvedTheme, theme, systemTheme, setTheme } = useTheme();
@@ -14,7 +13,7 @@ const NavBar = () => {
       <nav className="flex justify-between m-8 mt-5">
         <Logo className="flex self-center" currentTheme={currentTheme} />
         {/* <ColorSwitcher className="self-center" /> */}
-        <ColorSwitcher2 className=" self-center" />
+        <ThemeSwitcher className=" self-center" />
       </nav>
     </>
   );
